CHAD-17075: zigbee-smoke-detector lazy load subdrivers
