+++ drivers/SmartThings/zigbee-smoke-detector/src/aqara-gas/fingerprints.lua
@@ -1,0 +1,8 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local FINGERPRINTS = {
+    { mfr = "LUMI", model = "lumi.sensor_gas.acn02" }
+}
+
+return FINGERPRINTS
