+++ drivers/SmartThings/zigbee-smoke-detector/src/aqara-gas/init.lua
@@ -1,16 +1,6 @@
--- Copyright 2024 SmartThings
---
--- Licensed under the Apache License, Version 2.0 (the "License");
--- you may not use this file except in compliance with the License.
--- You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
+-- Copyright 2024 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
 local data_types = require "st.zigbee.data_types"
 local cluster_base = require "st.zigbee.cluster_base"
 local capabilities = require "st.capabilities"
@@ -31,9 +21,6 @@ local PRIVATE_LIFE_TIME_ATTRIBUTE_ID = 0x0128
 local PRIVATE_GAS_ZONE_STATUS_ATTRIBUTE_ID = 0x013A
 
 
-local FINGERPRINTS = {
-    { mfr = "LUMI", model = "lumi.sensor_gas.acn02" }
-}
 
 
 local CONFIGURATIONS = {
@@ -125,14 +112,6 @@ local function self_check_attr_handler(self, device, zone_status, zb_rx)
     PRIVATE_CLUSTER_ID, PRIVATE_SELF_CHECK_ATTRIBUTE_ID, MFG_CODE, data_types.Boolean, true))
 end
 
-local function is_aqara_products(opts, driver, device)
-  for _, fingerprint in ipairs(FINGERPRINTS) do
-    if device:get_manufacturer() == fingerprint.mfr and device:get_model() == fingerprint.model then
-      return true
-    end
-  end
-  return false
-end
 
 local function device_init(driver, device)
   if CONFIGURATIONS ~= nil then
@@ -186,8 +165,7 @@ local aqara_gas_detector_handler = {
       [startSelfCheckCommandName] = self_check_attr_handler
     },
   },
-  can_handle = is_aqara_products
+  can_handle = require("aqara-gas.can_handle"),
 }
 
 return aqara_gas_detector_handler
-
