+++ drivers/SmartThings/zigbee-smoke-detector/src/frient/can_handle.lua
@@ -1,0 +1,11 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local function frient_can_handle(opts, driver, device, ...)
+  if device:get_manufacturer() == "frient A/S" and (device:get_model() == "SMSZB-120" or device:get_model() == "HESZB-120") then
+    return true, require("frient")
+  end
+  return false
+end
+
+return frient_can_handle
